🔨Receipt selector looks good.
